+++ src/index.ts
@@ -1,4 +1,3 @@
 export type {Components, JSX} from './components';
-
-export * from './types/web-social-share-attributes';
-export * from './types/web-social-share-input';
+export * from './types/attributes';
+export * from './types/input';