+++ src/providers/twitter.ts
@@ -1,39 +1,30 @@
-import {WebSocialShareTwiterAttributes} from '../types/web-social-share-attributes';
+import {WebSocialShareTwiterAttributes} from '../types/attributes';
 
-import {isMobile, shareEncodedUrl, staticOpenNewWindow} from '../utils/utils';
+import {openNewWindow, shareEncodedUrl} from '../utils/utils';
 
-export const shareTwitter = async (attrs: WebSocialShareTwiterAttributes) => {
+export const shareTwitter = async ({
+  socialShareText,
+  socialShareVia,
+  socialShareHashtags,
+  socialShareUrl,
+  openWindowTarget: target
+}: WebSocialShareTwiterAttributes) => {
   let urlString = 'https://www.twitter.com/intent/tweet?';
 
-  if (attrs.socialShareText) {
-    urlString += 'text=' + encodeURIComponent(attrs.socialShareText);
+  if (socialShareText) {
+    urlString += 'text=' + encodeURIComponent(socialShareText);
   }
 
-  if (attrs.socialShareVia) {
-    urlString += '&via=' + encodeURIComponent(attrs.socialShareVia);
+  if (socialShareVia) {
+    urlString += '&via=' + encodeURIComponent(socialShareVia);
   }
 
-  if (attrs.socialShareHashtags) {
-    urlString += '&hashtags=' + encodeURIComponent(attrs.socialShareHashtags);
+  if (socialShareHashtags) {
+    urlString += '&hashtags=' + encodeURIComponent(socialShareHashtags);
   }
 
   //default to the current page if a URL isn't specified
-  urlString += '&url=' + shareEncodedUrl(attrs.socialShareUrl);
+  urlString += '&url=' + shareEncodedUrl(socialShareUrl);
 
-  if (isMobile()) {
-    staticOpenNewWindow(urlString);
-  } else {
-    window.open(
-      urlString,
-      'Twitter',
-      'toolbar=0,status=0,resizable=yes,width=' +
-        attrs.socialSharePopupWidth +
-        ',height=' +
-        attrs.socialSharePopupHeight +
-        ',top=' +
-        (window.innerHeight - attrs.socialSharePopupHeight) / 2 +
-        ',left=' +
-        (window.innerWidth - attrs.socialSharePopupWidth) / 2
-    );
-  }
+  openNewWindow({urlString, target});
 };